requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
SQLAlchemy
psycopg2-binary
